Adds in the HL7 ingestion into the CLI
